extracted game view from client view
