+++ pages/client.js
@@ -1,5 +1,6 @@
 const html = require('choo/html')
 const modal = require('../elements/web/modal')
+const gameView = require('../elements/web/game')
 
 var checkedForRecovery = false
 
@@ -141,33 +142,7 @@ module.exports = function (globalConfig) {
     if (state.client.connectivityState === globalConfig.connectivityStates.none) {
       send('setPage', 'INDEX')
     }
-
-    return html`
-<div>
-    <div class="row">
-        <h1>Welcome ${state.client.username}!</h1>
-    </div>
-    <div class="row">
-        <textarea name="code" id="code">${state.client.code ? state.client.code : ''}</textarea>
-        <button onclick=${sendCode}>send as code</button>
-    </div>
-    <div class="row">
-        <button onclick=${cleanExit}>clean Exit</button>
-    </div>
-</div>
-`
-
-    function sendCode (event) {
-      var code = document.getElementById('code').value
-      event.preventDefault()
-      if (!code || code.length === 0) return
-      send('client:sendCode', code)
-    }
-
-    function cleanExit (event) {
-      event.preventDefault()
-      send('client:quit', null)
-    }
+    return gameView(state, prev, send)
   }
 
 }
